#36 - feat : 테스트 패키지에 api/character 패키지 생성 및 관련 코드 해당 패키지로 이동
패키지를 분리하였다.
