--- core/data/src/test/java/com/hegunhee/maplefinder/data/TestParameter.kt
@@ -1,8 +1,0 @@
-package com.hegunhee.maplefinder.data
-
-object TestParameter {
-
-    const val CHARACTER_NAME = "엔버는함초롱"
-
-    val CHARACTER_OCID = "05a9d4c40c549bae73b529491cce27d5"
-}
