+++ core/data/src/test/java/com/hegunhee/maplefinder/data/api/TestParameter.kt
@@ -1,0 +1,8 @@
+package com.hegunhee.maplefinder.data.api
+
+object TestParameter {
+
+    const val CHARACTER_NAME = "엔버는함초롱"
+
+    const val CHARACTER_OCID = "05a9d4c40c549bae73b529491cce27d5"
+}
