+++ core/data/src/test/java/com/hegunhee/maplefinder/data/api/character/GetCharacterHyperStatUnitTest.kt
@@ -1,17 +1,16 @@
-package com.hegunhee.maplefinder.data.character
+package com.hegunhee.maplefinder.data.api.character
 
 import com.hegunhee.maplefinder.data.TestParameter
 import com.hegunhee.maplefinder.data.api.MapleCharacterApi
 import com.hegunhee.maplefinder.data.api.MapleOcidApi
 import com.hegunhee.maplefinder.data.getMapleApi
 import com.hegunhee.maplefinder.data.getMapleOcidApi
-import com.hegunhee.maplefinder.data.mapper.findMainStatName
 import com.hegunhee.maplefinder.data.mapper.toModel
 import kotlinx.coroutines.runBlocking
 import org.junit.Before
 import org.junit.Test
 
-class GetCharacterStatUnitTest {
+class GetCharacterHyperStatUnitTest {
 
     private lateinit var mapleOcidApi: MapleOcidApi
     private lateinit var mapleCharacterApi : MapleCharacterApi
@@ -26,20 +25,20 @@ class GetCharacterStatUnitTest {
     // 테스트 결과 캐릭터명 = 엔젤릭버스터
     // 새로운 규칙 : 2023-12-22 이후 ~ 어제의 날짜까지 조회 가능함
     @Test
-    fun `get character stat info test`() {
+    fun `get hyper stat info test`() {
         runBlocking {
             runCatching {
                 val ocid = mapleOcidApi.getOcid(characterName = TestParameter.CHARACTER_NAME).id
-                mapleCharacterApi.getCharacterStat(
+                mapleCharacterApi.getCharacterHyperStat(
                     ocid = ocid,
-                    date = "2024-02-20"
+                    date = "2024-01-22"
                 )
-            }.onSuccess { characterStat ->
-                if(characterStat.jobName == "엔젤릭버스터") {
-                    println(characterStat.toString())
+            }.onSuccess { characterHyperStat ->
+                if(characterHyperStat.jobName == "엔젤릭버스터") {
+                    println(characterHyperStat.toString())
                     assert(true)
                 }else {
-                    println(characterStat.toString())
+                    println(characterHyperStat.toString())
                     assert(false)
                 }
             }.onFailure {
@@ -50,42 +49,20 @@ class GetCharacterStatUnitTest {
     }
 
     @Test
-    fun `get character stat model test`() {
+    fun `get hyper stat model test`() {
         runBlocking {
             runCatching {
                 val ocid = mapleOcidApi.getOcid(characterName = TestParameter.CHARACTER_NAME).id
-                mapleCharacterApi.getCharacterStat(
+                mapleCharacterApi.getCharacterHyperStat(
                     ocid = ocid,
-                    date = "2024-02-20"
+                    date = "2024-01-22"
                 ).toModel()
-            }.onSuccess { characterStat ->
-                if(characterStat.jobName == "엔젤릭버스터") {
-                    println(characterStat.toString())
+            }.onSuccess { characterHyperStat ->
+                if(characterHyperStat.jobName == "엔젤릭버스터") {
+                    println(characterHyperStat.toString())
                     assert(true)
                 }else {
-                    println(characterStat.toString())
-                    assert(false)
-                }
-            }.onFailure {
-                println(it.message)
-                assert(false)
-            }
-        }
-    }
-
-    @Test
-    fun `find min sub stat test`() {
-        runBlocking {
-            runCatching {
-                val ocid = mapleOcidApi.getOcid(characterName = TestParameter.CHARACTER_NAME).id
-                mapleCharacterApi.getCharacterStat(
-                    ocid = ocid,
-                    date = "2024-02-20"
-                ).detailStatList.findMainStatName()
-            }.onSuccess { mainStat ->
-                if(mainStat == "덱스") {
-                    assert(true)
-                }else {
+                    println(characterHyperStat.toString())
                     assert(false)
                 }
             }.onFailure {
